[refeactor] - move struct's to the mta file
