--- mta/build-parameters.go
@@ -1,9 +1,0 @@
-package mta
-
-// BuildParameters - build params
-type BuildParameters struct {
-	Builder  string          `yaml:"builder,omitempty"`
-	Type     string          `yaml:"type,omitempty"`
-	Path     string          `yaml:"path,omitempty"`
-	Requires []BuildRequires `yaml:"requires,omitempty"`
-}
