--- mta/parameters.go
@@ -1,4 +1,0 @@
-package mta
-
-// Parameters - MTA parameters
-type Parameters map[string]interface{}
