--- mta/properties.go
@@ -1,4 +1,0 @@
-package mta
-
-// Properties - MTA map
-type Properties map[string]string
