--- mta/provides.go
@@ -1,7 +1,0 @@
-package mta
-
-// Provides - MTA struct
-type Provides struct {
-	Name       string
-	Properties Properties `yaml:"properties,omitempty"`
-}
