--- mta/resources.go
@@ -1,9 +1,0 @@
-package mta
-
-// Resources - resources section
-type Resources struct {
-	Name       string
-	Type       string
-	Parameters Parameters `yaml:"parameters,omitempty"`
-	Properties Properties `yaml:"properties,omitempty"`
-}
